requirements-ci.txt
boto3 ~= 1.26
pyhcl ~= 0.4
yamllint ~= 1.29
